+++ src/pumpwood_communication/microservice_abc/system/__init__.py
@@ -1,7 +1,8 @@
 """Module for pumpwood internals calls."""
 from .general import ABCSystemMicroservice
+from .permission import ABCPermissionMicroservice
 
 __docformat__ = "google"
 __all__ = [
-    ABCSystemMicroservice
+    ABCSystemMicroservice,ABCPermissionMicroservice
 ]